backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
httpx>=0.25.0
# psycopg2-binary==2.9.9  # uncomment for PostgreSQL
python-multipart==0.0.6
slowapi==0.1.9
pydub==0.25.1
python-dotenv==1.0.0
